+++ app/Http/Controllers/BlogsController.php
@@ -9,6 +9,16 @@ use App\BlogCategory;
 class BlogsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -97,6 +107,9 @@ class BlogsController extends Controller
      */
     public function edit($id)
     {
+        if (auth()->user()->super !== 1) {
+            return redirect('/dashboard')->with('error','Unauthorized Page');
+        }
         $blog = Blog::find($id);
         return view('blogs.edit')->with(['blog'=>$blog]);
     }
@@ -155,6 +168,9 @@ class BlogsController extends Controller
      */
     public function destroy($id)
     {
+        if (auth()->user()->super !== 1){
+            return redirect('/dashboard')->with('error','You can not delete this item');
+        }
         $blog = Blog::find($id);
 
         $blog->delete();
